capture-release-history/projects/01-release-history-tracker/requirements.txt
anthropic>=0.40.0
slack-sdk>=3.27.0
atlassian-python-api>=3.41.0
python-dotenv>=1.0.0
